Fixing a bug in setup
My last bug fix introduced a worse bug in the setup of DRAM. The main
problem was that the viral ref-seqs db would not match to its name
in the dram config dictionary, and so would fail to be written to
config. This will also fix an error where in some imported old (pre 1.4
dram) config names would not match to the new config.