AMG summary improved, some warnings prevented
This update changes the AMG summary so that it will provide all
information from both the Metabolic DB and The AMG DB. It is a bit of a
kitchen sink approach but we can always cut it down if needed and most
dram-v users I would call advanced.

There are also some small quality of life improvements.