Push a point release to react to ref-seqs changes
I have looked at the ref-seqs release that came out a little more than a
week ago 216 and I don't see any mention of one file. I am not sure why
the change took place. I will email the refseqs folks to make sure this
is not going back to a two file system in the next release. In the mean
time all the sequences that i would expect are in this one file based on
the catalog.
It is a small change so I am pushing this update without the full amount
of testing that we would sometimes do.
Thanks to @spencerlong1 and @nikolasbasler for noticing the change in
issue 254.