Replace sample with input_fasta for output
in DRAM 1, output (like raw-annotations) like had the column fasta
in DRAM 2 it was renamed to sample, which is incorrect since it isn't
the sample, it is the fasta. Reverting it to fasta like in DRAM 1
is a bit more logistically difficult because a large number of
nf processes (nf functions), and some python files use fasta
for the fastas file paths. So replacing sample with fasta
would cause collisions, and therefore we would need to first
replace all fasta keyworks with something like fasta_path
and then all sample with fasta. This route was easier
and probably less likely to introduce bugs.