Fix bug that wasn't running annotations on all input fasta
when porting to nf-core style, bug introduced where
annotations default sheet was treated as a channel when it wasn't before
so some processes would treat it as a single item channel and then
would only run the minimum of # of fasta or 1, which was 1
So only 1 fasta got ran.
Fixed this by changing it to not a channel, likt it was before the port.