Update modules code to include wave seqera container
using wave-cli, with cmd
for every module in modules/local, then adding the outputted url to
container outputer.url for the modules nextflow script, under the conda line.
This allows users to not just use conda, but also containers, and we don't have to build them
Ideally this would be added to a CI, but I haven't see where nf-core is doing that with a CI yet.