Rework distill to DRAM1 version
DRAM2's implementaiton of Distill utilized a sql db that was build up,
this db wasn't always the same as the raw annotaiton file, wasn't necessaryly faster
and added extra scheduelr jobs, so it was removed and repalced with the DRAM1 implementation.
The DRAM1 implementaiton takes the raw annotation file and does the calculation right
on that.

Removed the target count ids output as well since it was mostly used for the sql db distill
step.

Need to add back in the distill overview page that is from quast, ideally.

Removed pfam for now because of a bug of using mmseqs >=15 requires rebuilding profile db,
annotation code I believe was designed around compensating for the many to many
output that was being created (output file from mmseqs step with pfam was ~80-100 mb instead of
1/2 mb like it should have been), so remove pfam until next release when code is fixed and
maybe we can try out mseqs 17 even for faster queries.